Created CommerceUtil to manage trading
Part of #506

Changes:
- Reworked TradeUtil to be CommerceUtil.
- CommerceUtl uses Vehicle instead of Rover.
- Trade remove the 2 caches of potential trading settlement
- CommerceUtil internally manages a cache of potential deals
- Deal class holds the details of a potential trading deal
