--- mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/TradeUtil.java
@@ -1,807 +1,0 @@
-/*
- * Mars Simulation Project
- * TradeUtil.java
- * @date 2021-10-20
- * @author Scott Davis
- */
-package org.mars_sim.msp.core.person.ai.mission;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Set;
-import java.util.logging.Logger;
-
-import org.mars_sim.msp.core.Coordinates;
-import org.mars_sim.msp.core.Simulation;
-import org.mars_sim.msp.core.Unit;
-import org.mars_sim.msp.core.UnitManager;
-import org.mars_sim.msp.core.equipment.Container;
-import org.mars_sim.msp.core.equipment.ContainerUtil;
-import org.mars_sim.msp.core.equipment.EquipmentFactory;
-import org.mars_sim.msp.core.equipment.EquipmentType;
-import org.mars_sim.msp.core.goods.CreditManager;
-import org.mars_sim.msp.core.goods.Good;
-import org.mars_sim.msp.core.goods.GoodCategory;
-import org.mars_sim.msp.core.goods.GoodsManager;
-import org.mars_sim.msp.core.goods.GoodsUtil;
-import org.mars_sim.msp.core.person.PhysicalCondition;
-import org.mars_sim.msp.core.resource.AmountResource;
-import org.mars_sim.msp.core.resource.ItemResourceUtil;
-import org.mars_sim.msp.core.resource.Part;
-import org.mars_sim.msp.core.resource.ResourceUtil;
-import org.mars_sim.msp.core.structure.Settlement;
-import org.mars_sim.msp.core.time.MarsClock;
-import org.mars_sim.msp.core.vehicle.Rover;
-import org.mars_sim.msp.core.vehicle.Vehicle;
-import org.mars_sim.msp.core.vehicle.VehicleType;
-
-/**
- * Utility class for static methods for Trade Mission. TODO externalize strings
- */
-public final class TradeUtil {
-
-	/** default logger. */
-	private static final Logger logger = Logger.getLogger(TradeUtil.class.getName());
-
-	/**
-	 * Credit limit under which a seller is willing to sell goods to a buyer. Buyer
-	 * must pay off credit to under limit to continue buying.
-	 */
-	public static final double SELL_CREDIT_LIMIT = 10_000_000D;
-
-	/** Estimated mission parts mass. */
-	private static final double MISSION_BASE_MASS = 2_000D;
-
-	/** Minimum mass (kg) of life support resources to leave at settlement. */
-	private static final int MIN_LIFE_SUPPORT_RESOURCES = 100;
-
-	/** Minimum number of repair parts to leave at settlement. */
-	private static final int MIN_REPAIR_PARTS = 20;
-
-	/** Minimum number of repair parts to leave at settlement. */
-	private static final int MIN_NUM_EQUIPMENT = 10;
-
-	/** Cache for the best trade settlement. */
-	public static Settlement bestTradeSettlementCache = null;
-
-	private static final int OXYGEN_ID = ResourceUtil.oxygenID;
-	private static final int WATER_ID = ResourceUtil.waterID;
-	private static final int FOOD_ID = ResourceUtil.foodID;
-
-	private static Simulation sim = Simulation.instance();
-	private static MissionManager missionManager = sim.getMissionManager();
-	private static UnitManager unitManager = sim.getUnitManager();
-
-			
-	/**
-	 * Private constructor for utility class.
-	 */
-	private TradeUtil() {
-	};
-
-	/**
-	 * Gets the best trade value for a given settlement.
-	 * 
-	 * @param startingSettlement the settlement to trade from.
-	 * @param rover              the rover to carry the trade.
-	 * @return the best value (value points) for trade.
-	 * @throws Exception if error while getting best trade profit.
-	 */
-	public static double getBestTradeProfit(Settlement startingSettlement, Rover rover) {
-		double bestProfit = 0D;
-		Settlement bestSettlement = null;
-
-		for (Settlement tradingSettlement : unitManager.getSettlements()) {
-			if (tradingSettlement != startingSettlement && tradingSettlement.isMissionEnable(MissionType.TRADE)) {
-
-				boolean hasCurrentTradeMission = hasCurrentTradeMission(startingSettlement, tradingSettlement);
-
-				double settlementRange = Coordinates.computeDistance(tradingSettlement.getCoordinates(), startingSettlement.getCoordinates());
-				boolean withinRange = (settlementRange <= (rover.getRange(MissionType.TRADE) * .8D));
-
-				if (!hasCurrentTradeMission && withinRange) {
-					// double startTime = System.currentTimeMillis();
-
-					double profit = getEstimatedTradeProfit(startingSettlement, rover, tradingSettlement);
-					// double endTime = System.currentTimeMillis();
-//					 logger.finest("getEstimatedTradeProfit " + (endTime - startTime));
-					if (profit > bestProfit) {
-						bestProfit = profit;
-						bestSettlement = tradingSettlement;
-					}
-				}
-			}
-		}
-
-		// Set settlement cache.
-		bestTradeSettlementCache = bestSettlement;
-
-		return bestProfit;
-	}
-
-	/**
-	 * Checks if there is currently a trade mission between two settlements.
-	 * 
-	 * @param settlement1 the first settlement.
-	 * @param settlement2 the second settlement.
-	 * @return true if current trade mission between settlements.
-	 */
-	private static boolean hasCurrentTradeMission(Settlement settlement1, Settlement settlement2) {
-		boolean result = false;
-
-		// MissionManager manager = Simulation.instance().getMissionManager();
-		Iterator<Mission> i = missionManager.getMissions().iterator();
-		while (i.hasNext()) {
-			Mission mission = i.next();
-			if (mission instanceof Trade) {
-				Trade tradeMission = (Trade) mission;
-				Settlement startingSettlement = tradeMission.getStartingSettlement();
-				Settlement tradingSettlement = tradeMission.getTradingSettlement();
-				if (startingSettlement.equals(settlement1) && tradingSettlement.equals(settlement2)) {
-					result = true;
-					break;
-				} else if (startingSettlement.equals(settlement2) && tradingSettlement.equals(settlement1)) {
-					result = true;
-					break;
-				}
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Gets the estimated trade profit from one settlement to another.
-	 * 
-	 * @param startingSettlement the settlement to trade from.
-	 * @param rover              the rover to carry the trade goods.
-	 * @param tradingSettlement  the settlement to trade to.
-	 * @return the trade profit (value points)
-	 * @throws Exception if error getting the estimated trade profit.
-	 */
-	private static double getEstimatedTradeProfit(Settlement startingSettlement, Rover rover,
-			Settlement tradingSettlement) {
-
-		// Determine estimated trade revenue.
-		double revenue = getEstimatedTradeRevenue(startingSettlement, rover, tradingSettlement);
-
-		// Determine estimated mission cost.
-		double distance = startingSettlement.getCoordinates().getDistance(tradingSettlement.getCoordinates()) * 2D;
-		double cost = getEstimatedMissionCost(startingSettlement, rover, distance);
-
-		return revenue - cost;
-	}
-
-	/**
-	 * Gets the estimated trade revenue from one settlement to another.
-	 * 
-	 * @param startingSettlement the settlement to trade from.
-	 * @param rover              the rover to carry the trade goods.
-	 * @param tradingSettlement  the settlement to trade to.
-	 * @return the trade revenue (value points).
-	 * @throws Exception if error getting the estimated trade revenue.
-	 */
-	private static double getEstimatedTradeRevenue(Settlement startingSettlement, Rover rover,
-			Settlement tradingSettlement) {
-
-		// Get credit between starting settlement and trading settlement.
-		double credit = CreditManager.getCredit(startingSettlement, tradingSettlement);
-
-		Map<Good, Integer> buyLoad = null;
-		if (credit > (TradeUtil.SELL_CREDIT_LIMIT * -1D)) {
-			// Determine desired buy load,
-			buyLoad = TradeUtil.getDesiredBuyLoad(startingSettlement, rover, tradingSettlement);
-		} else {
-			// Cannot buy from settlement due to credit limit.
-			buyLoad = new HashMap<Good, Integer>(0);
-		}
-
-		Map<Good, Integer> sellLoad = null;
-		if (credit < TradeUtil.SELL_CREDIT_LIMIT) {
-			// Determine sell load.
-			sellLoad = TradeUtil.determineBestSellLoad(startingSettlement, rover, tradingSettlement);
-		} else {
-			// Will not sell to settlement due to credit limit.
-			sellLoad = new HashMap<Good, Integer>(0);
-		}
-
-		double sellingCreditHome = TradeUtil.determineLoadCredit(sellLoad, startingSettlement, false);
-		double sellingCreditRemote = TradeUtil.determineLoadCredit(sellLoad, tradingSettlement, true);
-		double sellingProfit = sellingCreditRemote - sellingCreditHome;
-
-		double buyingCreditHome = TradeUtil.determineLoadCredit(buyLoad, startingSettlement, true);
-		double buyingCreditRemote = TradeUtil.determineLoadCredit(buyLoad, tradingSettlement, false);
-		double buyingProfit = buyingCreditHome - buyingCreditRemote;
-
-		double totalProfit = sellingProfit + buyingProfit;
-
-		return totalProfit;
-	}
-
-	/**
-	 * Gets the desired buy load from a trading settlement.
-	 * 
-	 * @param startingSettlement the settlement that is buying.
-	 * @param rover              the rover used for trade.
-	 * @param tradingSettlement  the settlement to buy from.
-	 * @return the desired buy load.
-	 * @throws Exception if error determining the buy load.
-	 */
-	public static Map<Good, Integer> getDesiredBuyLoad(Settlement startingSettlement, Rover rover,
-			Settlement tradingSettlement) {
-
-		// Determine best buy load.
-		Map<Good, Integer> buyLoad = determineLoad(startingSettlement, tradingSettlement, rover,
-				Double.POSITIVE_INFINITY);
-
-		return buyLoad;
-	}
-
-	/**
-	 * Determines the best sell load from a settlement to another.
-	 * 
-	 * @param startingSettlement the settlement to trade from.
-	 * @param rover              the rover to carry the goods.
-	 * @param tradingSettlement  the settlement to trade to.
-	 * @return a map of goods and numbers in the load.
-	 * @throws Exception if error determining best sell load.
-	 */
-	static Map<Good, Integer> determineBestSellLoad(Settlement startingSettlement, Rover rover,
-			Settlement tradingSettlement) {
-
-		// Determine best sell load.
-		Map<Good, Integer> sellLoad = determineLoad(tradingSettlement, startingSettlement, rover,
-				Double.POSITIVE_INFINITY);
-
-		return sellLoad;
-	}
-
-	/**
-	 * Determines the load between a buying settlement and a selling settlement.
-	 * 
-	 * @param buyingSettlement  the settlement buying the goods.
-	 * @param sellingSettlement the settlement selling the goods.
-	 * @param rover             the rover to carry the goods.
-	 * @param maxBuyValue       the maximum value the selling settlement will
-	 *                          permit.
-	 * @return map of goods and their number.
-	 * @throws Exception if error determining the load.
-	 */
-	public static Map<Good, Integer> determineLoad(Settlement buyingSettlement, Settlement sellingSettlement,
-			Rover rover, double maxBuyValue) {
-
-		Map<Good, Integer> tradeList = new HashMap<Good, Integer>();
-		boolean hasRover = false;
-		GoodsManager buyerGoodsManager = buyingSettlement.getGoodsManager();
-		buyerGoodsManager.prepareForLoadCalculation();
-		GoodsManager sellerGoodsManager = sellingSettlement.getGoodsManager();
-		sellerGoodsManager.prepareForLoadCalculation();
-
-		double massCapacity = rover.getCargoCapacity();
-
-		// Subtract mission base mass (estimated).
-		double missionPartsMass = MISSION_BASE_MASS;
-		if (massCapacity < missionPartsMass)
-			missionPartsMass = massCapacity;
-		massCapacity -= missionPartsMass;
-
-		// Determine repair parts for trip.
-		Set<Integer> repairParts = rover.getMalfunctionManager().getRepairPartProbabilities().keySet();
-
-		// Determine the load.
-		boolean done = false;
-		double buyerLoadValue = 0D;
-		Good previousGood = null;
-		Set<Good> nonTradeGoods = Collections.emptySet();
-		while (!done) {
-			double remainingBuyValue = maxBuyValue - buyerLoadValue;
-			Good good = findBestTradeGood(sellingSettlement, buyingSettlement, tradeList, nonTradeGoods, massCapacity,
-					hasRover, rover, previousGood, false, repairParts, remainingBuyValue);
-			if (good != null) {
-				try {
-					boolean isAmountResource = good.getCategory() == GoodCategory.AMOUNT_RESOURCE;
-					boolean isItemResource = good.getCategory() == GoodCategory.ITEM_RESOURCE;
-					AmountResource resource = null;
-					
-					// Add resource container if needed.
-					if (isAmountResource) {
-						resource = ResourceUtil.findAmountResource(good.getID());
-						Container container = getAvailableContainerForResource(resource,
-								sellingSettlement, tradeList);
-						if (container != null) {
-							Good containerGood = GoodsUtil.getEquipmentGood(container.getEquipmentType());
-							massCapacity -= container.getBaseMass();
-							int containerNum = 0;
-							if (tradeList.containsKey(containerGood))
-								containerNum = tradeList.get(containerGood);
-							double containerSupply = containerGood.getNumberForSettlement(buyingSettlement);
-							double totalContainerNum = containerNum + containerSupply;
-							buyerLoadValue += buyerGoodsManager.determineGoodValueWithSupply(containerGood, totalContainerNum);
-							tradeList.put(containerGood, (containerNum + 1));
-						} else
-							logger.warning("container for " + resource.getName() + " not available.");
-					}
-
-					int itemResourceNum = 0;
-					if (isItemResource) {
-						itemResourceNum = getNumItemResourcesToTrade(good, sellingSettlement, buyingSettlement,
-								tradeList, massCapacity, remainingBuyValue);
-					}
-
-					// Add good.
-					if (good.getCategory() == GoodCategory.VEHICLE)
-						hasRover = true;
-					else {
-						int number = 1;
-						if (isAmountResource)
-							number = (int) getResourceTradeAmount(resource);
-						else if (isItemResource)
-							number = itemResourceNum;
-						massCapacity -= (good.getMassPerItem() * number);
-					}
-					
-					int currentNum = 0;
-					if (tradeList.containsKey(good))
-						currentNum = tradeList.get(good);
-					double supply = good.getNumberForSettlement(buyingSettlement);
-					double goodNum = 1D;
-					
-					if (isAmountResource)
-						goodNum = getResourceTradeAmount(resource);
-					if (isItemResource)
-						goodNum = itemResourceNum;
-					
-					double buyGoodValue = buyerGoodsManager.determineGoodValueWithSupply(good, (supply + currentNum + goodNum));
-					
-					if (isAmountResource) {
-						double tradeAmount = getResourceTradeAmount(resource);
-						buyGoodValue *= tradeAmount;
-					}
-					if (isItemResource) {
-						buyGoodValue *= itemResourceNum;
-					}
-					buyerLoadValue += buyGoodValue;
-					int newNumber = currentNum + (int) goodNum;
-					tradeList.put(good, newNumber);
-				} catch (Exception e) {
-					done = true;
-				}
-			} else
-				done = true;
-
-			previousGood = good;
-		}
-
-		return tradeList;
-	}
-
-	/**
-	 * Determines the credit of a load to a settlement.
-	 * 
-	 * @param load       a map of the goods and their number.
-	 * @param settlement the settlement valuing the load.
-	 * @param buy        true if settlement is buying the load, false if selling.
-	 * @return credit of the load (value points * production cost).
-	 * @throws Exception if error determining the load credit.
-	 */
-	public static double determineLoadCredit(Map<Good, Integer> load, Settlement settlement, boolean buy) {
-		double result = 0D;
-
-		GoodsManager manager = settlement.getGoodsManager();
-
-		Iterator<Good> i = load.keySet().iterator();
-		while (i.hasNext()) {
-			Good good = i.next();
-			double cost = good.getCostOutput();
-			int goodNumber = load.get(good);
-			double supply = good.getNumberForSettlement(settlement);
-			double multiplier = 1D;
-			if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {	
-				double amount = getResourceTradeAmount(ResourceUtil.findAmountResource(good.getID()));
-				if (amount < 1) {
-					multiplier = 1;
-				}
-				else {
-					goodNumber /= (int) amount;
-					multiplier = amount;
-				}
-			}
-			else {	
-				multiplier = 1D;
-			}
-			
-			for (int x = 0; x < goodNumber; x++) {
-
-				double supplyAmount = 0D;
-				if (buy)
-					supplyAmount = supply + x;
-				else {
-					supplyAmount = supply - x;
-					if (supplyAmount < 0D)
-						supplyAmount = 0D;
-				}
-
-				double value = (manager.determineGoodValueWithSupply(good, supplyAmount) * multiplier);
-
-				result += value * cost;
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Finds the best trade good for a trade.
-	 * 
-	 * @param sellingSettlement the settlement selling the good.
-	 * @param buyingSettlement  the settlement buying the good.
-	 * @param tradedGoods       the map of goods traded so far.
-	 * @param nonTradeGoods     the set of goods not to trade.
-	 * @param remainingCapacity remaining general capacity (kg) in vehicle
-	 *                          inventory.
-	 * @param hasVehicle        true if a vehicle is in the trade goods.
-	 * @param missionRover      the rover carrying the goods.
-	 * @param previousGood      the previous trade good used in the trade.
-	 * @param allowNegValue     allow negative value goods.
-	 * @param repairParts       set of repair parts possibly needed for the trip.
-	 * @param maxBuyValue       the maximum value the item can be.
-	 * @return best good to trade or null if none found.
-	 * @throws Exception if error determining best trade good.
-	 */
-	private static Good findBestTradeGood(Settlement sellingSettlement, Settlement buyingSettlement,
-			Map<Good, Integer> tradedGoods, Set<Good> nonTradeGoods, double remainingCapacity, boolean hasVehicle,
-			Rover missionRover, Good previousGood, boolean allowNegValue, Set<Integer> repairParts,
-			double maxBuyValue) {
-
-		Good result = null;
-
-		// Check previous good first.
-		if (previousGood != null) {
-			double previousGoodValue = getTradeValue(previousGood, sellingSettlement, buyingSettlement, tradedGoods,
-					remainingCapacity, hasVehicle, missionRover, allowNegValue, repairParts);
-			if ((previousGoodValue > 0D) && (previousGoodValue < maxBuyValue))
-				result = previousGood;
-		}
-
-		// Check all goods.
-		if (result == null) {
-			double bestValue = 0D;
-			if (allowNegValue)
-				bestValue = Double.NEGATIVE_INFINITY;
-			Iterator<Good> i = buyingSettlement.getBuyList().iterator(); //GoodsUtil.getGoodsList();
-			while (i.hasNext()) {
-			Good good = i.next();
-				if (!nonTradeGoods.contains(good)) {
-					double tradeValue = getTradeValue(good, sellingSettlement, buyingSettlement, tradedGoods,
-							remainingCapacity, hasVehicle, missionRover, allowNegValue, repairParts);
-					if ((tradeValue > bestValue) && (tradeValue < maxBuyValue)) {
-						result = good;
-						bestValue = tradeValue;
-					}
-				}
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Gets the number of an item resource good that should be traded.
-	 * 
-	 * @param itemResourceGood  the item resource good.
-	 * @param sellingSettlement the settlement selling the good.
-	 * @param buyingSettlement  the settlement buying the good.
-	 * @param tradeList         the map of goods traded so far.
-	 * @param remainingCapacity remaining general capacity (kg) in vehicle
-	 *                          inventory.
-	 * @param maxBuyValue       the maximum buy value.
-	 * @return number of goods to trade.
-	 * @throws Exception if error determining number of goods.
-	 */
-	private static int getNumItemResourcesToTrade(Good itemResourceGood, Settlement sellingSettlement,
-			Settlement buyingSettlement, Map<Good, Integer> tradeList, double remainingCapacity, double maxBuyValue) {
-
-		int result = 0;
-
-		Part item = ItemResourceUtil.findItemResource(itemResourceGood.getID());
-
-		int sellingInventory = sellingSettlement.getItemResourceStored(item.getID());
-		int buyingInventory = buyingSettlement.getItemResourceStored(item.getID());
-
-		int numberTraded = 0;
-		if (tradeList.containsKey(itemResourceGood))
-			numberTraded = tradeList.get(itemResourceGood);
-
-		int roverLimit = (int) (remainingCapacity / item.getMassPerItem());
-
-		int totalTraded = numberTraded;
-		double totalBuyingValue = 0D;
-		boolean limitReached = false;
-		while (!limitReached) {
-
-			double sellingSupplyAmount = sellingInventory - totalTraded - 1D;
-			double sellingValue = sellingSettlement.getGoodsManager().determineGoodValueWithSupply(itemResourceGood,
-					sellingSupplyAmount);
-			double buyingSupplyAmount = buyingInventory + totalTraded + 1D;
-			double buyingValue = buyingSettlement.getGoodsManager().determineGoodValueWithSupply(itemResourceGood,
-					buyingSupplyAmount);
-
-			if (buyingValue <= sellingValue)
-				limitReached = true;
-			if (totalTraded + 1 > sellingInventory)
-				limitReached = true;
-			if (totalTraded + 1 > roverLimit)
-				limitReached = true;
-			if ((totalBuyingValue + buyingValue) >= maxBuyValue)
-				limitReached = true;
-
-			if (!limitReached) {
-				result++;
-				totalTraded = numberTraded + result;
-				totalBuyingValue += buyingValue;
-			}
-		}
-
-		// Result shouldn't be zero, but just in case it is.
-		if (result == 0)
-			result = 1;
-		return result;
-	}
-
-	/**
-	 * Gets the trade value of a good.
-	 * 
-	 * @param good              the good
-	 * @param sellingSettlement the settlement selling the good.
-	 * @param buyingSettlement  the settlement buying the good.
-	 * @param tradedGoods       the map of goods traded so far.
-	 * @param remainingCapacity remaining general capacity (kg) in vehicle
-	 *                          inventory.
-	 * @param hasVehicle        true if a vehicle is in the trade goods.
-	 * @param missionRover      the rover carrying the goods.
-	 * @param allowNegValue     allow negative value goods.
-	 * @param repairParts       set of repair parts possibly needed for the trip.
-	 * @return trade value of good.
-	 * @throws Exception if error determining trade value.
-	 */
-	private static double getTradeValue(Good good, Settlement sellingSettlement, Settlement buyingSettlement,
-			Map<Good, Integer> tradedGoods, double remainingCapacity, boolean hasVehicle, Rover missionRover,
-			boolean allowNegValue, Set<Integer> repairParts) {
-
-		double result = Double.NEGATIVE_INFINITY;
-		AmountResource resource = null;
-		double amountTraded = 0D;
-		if (tradedGoods.containsKey(good))
-			amountTraded += tradedGoods.get(good).doubleValue();
-
-		double sellingInventory = getNumInInventory(good, sellingSettlement);
-		double sellingSupplyAmount = sellingInventory - amountTraded - 1D;
-		if (sellingSupplyAmount < 0D)
-			sellingSupplyAmount = 0D;
-		double sellingValue = sellingSettlement.getGoodsManager().determineGoodValueWithSupply(good, sellingSupplyAmount);
-		if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {
-			resource = ResourceUtil.findAmountResource(good.getID());
-			sellingValue *= getResourceTradeAmount(resource);
-		}
-		boolean allTraded = (sellingInventory <= amountTraded);
-
-		double buyingInventory = getNumInInventory(good, buyingSettlement);
-		double buyingSupplyAmount = buyingInventory + amountTraded + 1D;
-		if (buyingSupplyAmount < 0D)
-			buyingSupplyAmount = 0D;
-		double buyingValue = buyingSettlement.getGoodsManager().determineGoodValueWithSupply(good, buyingSupplyAmount);
-		if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE)
-			buyingValue *= getResourceTradeAmount(resource);
-
-		boolean profitable = (buyingValue > sellingValue);
-		boolean hasBuyValue = buyingValue > 0D;
-		if ((allowNegValue || profitable) && hasBuyValue && !allTraded) {
-			// Check if rover inventory has capacity for the good.
-			boolean isRoverCapacity = hasCapacityInInventory(good, buyingSettlement, remainingCapacity, hasVehicle);
-
-			boolean isContainerAvailable = true;
-			if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {
-				Container container = getAvailableContainerForResource(resource,
-						sellingSettlement, tradedGoods);
-				isContainerAvailable = (container != null);
-			}
-
-			boolean isMissionRover = false;
-			if (good.getCategory() == GoodCategory.VEHICLE) {
-				if (good.getName().equalsIgnoreCase(missionRover.getDescription())) {
-					if (sellingInventory == 1D)
-						isMissionRover = true;
-				}
-			}
-
-			boolean enoughResourceForContainer = true;
-			if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {
-				enoughResourceForContainer = (sellingSupplyAmount >= getResourceTradeAmount(resource));
-			}
-
-			boolean enoughEVASuits = true;
-			boolean enoughEquipment = true;
-			if (good.getCategory() == GoodCategory.EQUIPMENT
-					|| good.getCategory() == GoodCategory.CONTAINER) {	
-				if (good.getEquipmentType() == EquipmentType.EVA_SUIT) {
-					double remainingSuits = sellingInventory - amountTraded;
-					int requiredSuits = Trade.MAX_MEMBERS + 2;
-					enoughEVASuits = remainingSuits > requiredSuits;
-				}
-				else {
-					double remaining = sellingInventory - amountTraded;
-					enoughEquipment = remaining > MIN_NUM_EQUIPMENT;
-				}
-
-			}
-
-			boolean enoughRepairParts = true;
-			if (good.getCategory() == GoodCategory.ITEM_RESOURCE) {
-				if (repairParts.contains(good.getID())) {
-					if (sellingSupplyAmount < MIN_REPAIR_PARTS)
-						enoughRepairParts = false;
-				}
-			}
-
-			boolean enoughLifeSupportResources = true;
-			if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {
-				if (resource.isLifeSupport() && sellingSupplyAmount < MIN_LIFE_SUPPORT_RESOURCES)
-					enoughLifeSupportResources = false;
-			}
-
-			if (isRoverCapacity && isContainerAvailable && !isMissionRover && enoughResourceForContainer
-					&& enoughEVASuits && enoughEquipment && enoughRepairParts && enoughLifeSupportResources) {
-				result = buyingValue - sellingValue;
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Checks if capacity in inventory for good.
-	 * 
-	 * @param good              the good to check for.
-	 * @param remainingCapacity remaining general capacity (kg) in vehicle
-	 *                          inventory.
-	 * @param hasVehicle        true if good load already includes a vehicle.
-	 * @return true if capacity for good.
-	 * @throws Exception if error checking for capacity.
-	 */
-	private static boolean hasCapacityInInventory(Good good, Settlement settlement, double remainingCapacity, boolean hasVehicle) {
-		boolean result = false;
-		if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {
-			result = (remainingCapacity >= getResourceTradeAmount(ResourceUtil.findAmountResource(good.getID())));
-		} else if (good.getCategory() == GoodCategory.ITEM_RESOURCE)
-			result = remainingCapacity >= ItemResourceUtil.findItemResource(good.getID()).getMassPerItem();
-		else if (good.getCategory() == GoodCategory.EQUIPMENT
-				|| good.getCategory() == GoodCategory.CONTAINER) {
-			result = (remainingCapacity >= EquipmentFactory.getEquipmentMass(good.getEquipmentType()));
-		} else if (good.getCategory() == GoodCategory.VEHICLE)
-			result = !hasVehicle;
-		return result;
-	}
-
-	/**
-	 * Gets the number of a good currently in the inventory.
-	 * 
-	 * @param good      the good to check.
-	 * @param inventory the inventory to check.
-	 * @return number of goods in inventory.
-	 * @throws Exception if error getting number of goods in inventory.
-	 */
-	public static double getNumInInventory(Good good, Settlement settlement) {
-		if (good.getCategory() == GoodCategory.AMOUNT_RESOURCE) {
-			return settlement.getAmountResourceStored(good.getID());
-		} else if (good.getCategory() == GoodCategory.ITEM_RESOURCE) {
-			return settlement.getItemResourceStored(good.getID());
-		} else if (good.getCategory() == GoodCategory.EQUIPMENT
-				|| good.getCategory() == GoodCategory.CONTAINER) {
-			return settlement.findNumEmptyContainersOfType(good.getEquipmentType(), false);
-		} else if (good.getCategory() == GoodCategory.VEHICLE) {
-			int count = 0;
-			VehicleType vehicleType = VehicleType.convertNameToVehicleType(good.getName());
-			Iterator<Unit> i = settlement.getVehicleTypeList(vehicleType).iterator();
-			while (i.hasNext()) {
-				Vehicle vehicle = (Vehicle) i.next();
-				boolean isEmpty = vehicle.isEmpty();
-				if (vehicle.getDescription().equalsIgnoreCase(good.getName()) && !vehicle.isReserved() && isEmpty) {
-					count++;
-				}
-			}
-			return count;
-		} else {
-			return 0D;
-		}
-	}
-
-	/**
-	 * Gets an available container for a given resource.
-	 * 
-	 * @param resource    the resource to check.
-	 * @param settlement  the settlement to check for containers.
-	 * @param tradedGoods the list of goods traded so far.
-	 * @return container for the resource or null if none.
-	 * @throws Exception if error.
-	 */
-	private static Container getAvailableContainerForResource(AmountResource resource, Settlement settlement,
-			Map<Good, Integer> tradedGoods) {
-
-		Container result = null;
-
-		EquipmentType containerType = ContainerUtil.getContainerTypeNeeded(resource.getPhase());
-
-		int containersStored = settlement.findNumEmptyContainersOfType(containerType, false);
-
-		Good containerGood = GoodsUtil.getEquipmentGood(containerType);
-		int containersTraded = 0;
-		if (tradedGoods.containsKey(containerGood))
-			containersTraded = tradedGoods.get(containerGood);
-
-		if (containersStored > containersTraded)
-			result = settlement.findContainer(containerType, true, resource.getID());
-
-		return result;
-	}
-
-	/**
-	 * Gets the estimated trade mission cost.
-	 * 
-	 * @param startingSettlement the settlement starting the trade mission.
-	 * @param rover              the mission rover.
-	 * @param distance           the distance (km) of the mission trip.
-	 * @return the cost of the mission (value points).
-	 * @throws Exception if error getting the estimated mission cost.
-	 */
-	public static double getEstimatedMissionCost(Settlement startingSettlement, Rover rover, double distance) {
-		Map<Good, Integer> neededResources = new HashMap<Good, Integer>(4);
-
-		// Get required fuel.
-		Good fuelGood = GoodsUtil.getGood(rover.getFuelType());
-		neededResources.put(fuelGood, (int) VehicleMission.getFuelNeededForTrip(rover, distance, 
-				(rover.getCumFuelEconomy() + rover.getEstimatedFuelEconomy()) / VehicleMission.FE_FACTOR, false));
-
-		// Get estimated trip time.
-		double averageSpeed = rover.getBaseSpeed() / 2D;
-		double averageSpeedMillisol = averageSpeed / MarsClock.convertSecondsToMillisols(60D * 60D);
-		double tripTimeSols = ((distance / averageSpeedMillisol) + 1000D) / 1000D;
-
-		double life_support_margin = Vehicle.getLifeSupportRangeErrorMargin();
-		// Get oxygen amount.
-		double oxygenAmount = PhysicalCondition.getOxygenConsumptionRate() * tripTimeSols * Trade.MAX_MEMBERS
-				* Mission.OXYGEN_MARGIN * life_support_margin;
-		Good oxygenGood = GoodsUtil.getGood(OXYGEN_ID);
-		neededResources.put(oxygenGood, (int) oxygenAmount);
-
-		// Get water amount.
-		double waterAmount = PhysicalCondition.getWaterConsumptionRate() * tripTimeSols * Trade.MAX_MEMBERS
-				* Mission.WATER_MARGIN * life_support_margin;
-		Good waterGood = GoodsUtil.getGood(WATER_ID);
-		neededResources.put(waterGood, (int) waterAmount);
-
-		// Get food amount.
-		double foodAmount = PhysicalCondition.getFoodConsumptionRate() * tripTimeSols * Trade.MAX_MEMBERS
-				* Mission.FOOD_MARGIN * life_support_margin;
-		Good foodGood = GoodsUtil.getGood(FOOD_ID);
-		neededResources.put(foodGood, (int) foodAmount);
-
-		// Get cost of resources.
-		return determineLoadCredit(neededResources, startingSettlement, false);
-	}
-	
-	/**
-	 * Gets the amount of a resource that should be traded based on its standard
-	 * container capacity.
-	 * 
-	 * @param resource the amount resource.
-	 * @return amount (kg) of resource to trade.
-	 * @throws Exception if error determining container.
-	 */
-	private static double getResourceTradeAmount(AmountResource resource) {
-		EquipmentType containerType = ContainerUtil.getContainerTypeNeeded(resource.getPhase());
-		return ContainerUtil.getContainerCapacity(containerType);
-	}
-	
-}
